'-p' flag was given assuming it was 'progress', it isn't
The '-p' flag indicates that the file permissions must be preserved.
I am fairly certain that this flag is NOT needed, furthermore, it
would only be enabled when running in `--verbose` mode, which is
insane if such an important flag was actually required.

The '--progress' flag indicates that more progress information is
shown.  That is the flag that should have been used from the start.